requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==22.6.0
idna==3.6
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
